requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.2.1
Django==3.1.4
djangorestframework==3.12.2
jwt==1.1.0
mysqlclient==2.0.1
pycparser==2.20
pytz==2020.5
simplejwt==2.0.1
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
